Create IndividualProjAnalysis.R
This code is for education and economic data from the World Bank database (see the python code for the method used to acquire the data).  It includes linear and binary regression analysis.  The visualizations were completed solely on the data that included all years.  It includes histograms of the Gini index, poverty percentage, and primary and secondary education percentages.  In addition there are two scatterplots, one each for the Gini index and for the poverty percentage.  They include data points for secondary and primary education by gender (a total of four datasets).