Update description
This code gets data from the World Bank databases through their API.  The data is then averaged over the years collected and loaded into a two dimensional database (indicators vs country).  The analysis for this data is on education versus economic equality.  If a country was missing either education data or economic data it was dropped from the dataset.  Additionally a handful of binary variables were created to assist in the analysis.  The datasets are then saved to csv files.